Rename NewUpload view with View postfix
